Move `Render` to a private submodule
